Finish lec-12

+++ app/Http/Controllers/ProductController.php
@@ -6,7 +6,7 @@ use App\Models\Product;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Relations\Relation;
 use Illuminate\Http\RedirectResponse;
-use Illuminate\Support\Facades\Route;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\View\View;
 use Psr\Http\Message\ServerRequestInterface;
 
@@ -78,6 +78,8 @@ class ProductController extends SearchableController
 
     function list(ServerRequestInterface $request): View
     {
+        Gate::authorize('list', Product::class);
+
         $criteria = $this->prepareCriteria($request->getQueryParams());
         $query = $this
             ->search($criteria)
@@ -94,6 +96,8 @@ class ProductController extends SearchableController
     {
         $product = $this->find($productCode);
 
+        Gate::authorize('view', $product);
+
         return view('products.view', [
             'product' => $product,
         ]);
@@ -101,6 +105,8 @@ class ProductController extends SearchableController
 
     function showCreateForm(CategoryController $categoryController): View
     {
+        Gate::authorize('create', Product::class);
+
         $categories = $categoryController->getQuery()->get();
 
         return view('products.create-form', [
@@ -112,6 +118,8 @@ class ProductController extends SearchableController
         ServerRequestInterface $request,
         CategoryController $categoryController,
     ): RedirectResponse {
+        Gate::authorize('create', Product::class);
+
         $data = $request->getParsedBody();
         $category = $categoryController->find($data['category']);
 
@@ -131,6 +139,9 @@ class ProductController extends SearchableController
         string $productCode,
     ): View {
         $product = $this->find($productCode);
+
+        Gate::authorize('update', $product);
+
         $categories = $categoryController->getQuery()->get();
 
         return view('products.update-form', [
@@ -144,10 +155,13 @@ class ProductController extends SearchableController
         CategoryController $categoryController,
         string $productCode,
     ): RedirectResponse {
+        $product = $this->find($productCode);
+
+        Gate::authorize('update', $product);
+
         $data = $request->getParsedBody();
         $category = $categoryController->find($data['category']);
 
-        $product = $this->find($productCode);
         $product->fill($data);
         $product->category()->associate($category);
         $product->save();
@@ -162,6 +176,9 @@ class ProductController extends SearchableController
     function delete(string $productCode): RedirectResponse
     {
         $product = $this->find($productCode);
+
+        Gate::authorize('create', $product);
+
         $product->delete();
 
         return redirect(
@@ -176,6 +193,9 @@ class ProductController extends SearchableController
         string $productCode,
     ): View {
         $product = $this->find($productCode);
+
+        Gate::authorize('view', $product);
+
         $criteria = $shopController->prepareCriteria($request->getQueryParams());
         $query = $shopController
             ->filter($product->shops(), $criteria)
@@ -194,6 +214,9 @@ class ProductController extends SearchableController
         string $productCode,
     ): View {
         $product = $this->find($productCode);
+
+        Gate::authorize('update', $product);
+
         $query = $shopController
             ->getQuery()
             ->whereDoesntHave(
@@ -221,6 +244,9 @@ class ProductController extends SearchableController
         string $productCode,
     ): RedirectResponse {
         $product = $this->find($productCode);
+
+        Gate::authorize('update', $product);
+
         $data = $request->getParsedBody();
 
         $shop = $shopController
@@ -246,6 +272,9 @@ class ProductController extends SearchableController
         string $productCode,
     ): RedirectResponse {
         $product = $this->find($productCode);
+
+        Gate::authorize('update', $product);
+
         $data = $request->getParsedBody();
 
         $shop = $product->shops()
